+++ server/src-test/Data/Parser/JSONPathSpec.hs
@@ -3,35 +3,41 @@ module Data.Parser.JSONPathSpec (spec) where
 import Data.Aeson.Key qualified as K
 import Data.Parser.JSONPath
 import Data.Text qualified as T
-import Hasura.Base.Error (encodeJSONPath)
 import Hasura.Prelude
 import Test.Hspec
 import Test.QuickCheck
 
 spec :: Spec
-spec = describe "encode and parse JSONPath" $ do
-  it "JSONPath encoder" $
-    forM_ generateTestEncodeJSONPath $ \(jsonPath, result) ->
-      encodeJSONPath jsonPath `shouldBe` result
+spec = do
+  describe "encoding a JSON path" $ do
+    it "encodes a one-level path" $
+      encodeJSONPath [Key "ABCD"] `shouldBe` "$.ABCD"
 
-  describe "JSONPath parser" $ do
-    it "Single $" $
-      parseJSONPath "$" `shouldBe` (Right [] :: Either String JSONPath)
+    it "encodes a multi-level path" $
+      encodeJSONPath [Key "7seven", Index 0, Key "@!^@*#(!("] `shouldBe` "$[\"7seven\"][0][\"@!^@*#(!(\"]"
 
-    it "Random json paths" $
+    it "escapes control characters and quotes" $
+      encodeJSONPath [Key "/\\ '\" \t\r\n \xfffd"] `shouldBe` "$[\"/\\\\ '\\\" \\t\\r\\n \xfffd\"]"
+
+  describe "parsing a JSON path" $ do
+    it "parses a single '$'" $
+      parseJSONPath "$" `shouldBe` Right []
+
+    it "parses bracketed single quotes" $
+      parseJSONPath "$['foo \\' \" bar']" `shouldBe` Right [Key "foo ' \" bar"]
+
+    it "parses bracketed double quotes" $
+      parseJSONPath "$[\"bar ' \\\" foo\"]" `shouldBe` Right [Key "bar ' \" foo"]
+
+  describe "the round trip" $ do
+    it "encodes and parses random JSON paths" $
       withMaxSuccess 1000 $
         forAll (resize 20 generateJSONPath) $ \jsonPath ->
           let encPath = encodeJSONPath jsonPath
-              parsedJSONPathE = parseJSONPath $ T.pack encPath
+              parsedJSONPathE = parseJSONPath encPath
            in case parsedJSONPathE of
-                Left err -> counterexample (err <> ": " <> encPath) False
-                Right parsedJSONPath -> property $ parsedJSONPath == jsonPath
-
-generateTestEncodeJSONPath :: [(JSONPath, String)]
-generateTestEncodeJSONPath =
-  [ ([Key "7seven", Index 0, Key "@!^@*#(!("], "$['7seven'][0]['@!^@*#(!(']"),
-    ([Key "ABCD"], "$.ABCD")
-  ]
+                Left err -> counterexample (T.unpack (err <> ": " <> encPath)) False
+                Right parsedJSONPath -> property $ parsedJSONPath === jsonPath
 
 generateJSONPath :: Gen JSONPath
 generateJSONPath = map (either id id) <$> listOf1 genPathElementEither
@@ -41,4 +47,4 @@ generateJSONPath = map (either id id) <$> listOf1 genPathElementEither
       keyRight <- Right <$> genKey
       elements [indexLeft, keyRight]
     genIndex = Index <$> choose (0, 100)
-    genKey = Key . K.fromText . T.pack <$> listOf1 (elements $ alphaNumerics ++ ".,!@#$%^&*_-?:;|/\"")
+    genKey = Key . K.fromText . T.pack <$> listOf1 arbitraryUnicodeChar